Revamped sliding window approach
Previously after each window slide, the new window was taken from the unmodified audio.
Now a window/2 size slice is where the subtraction happens. After all notes are removed, the first w/2 is discarded, but the end of the subtraction that happend in the second half of the window is preserved. The new half-window slices is than attached to the end.